requirements.txt
# Основные зависимости
streamlit>=1.32.0
python-dotenv>=1.0.0
requests>=2.31.0

# Работа с данными
numpy>=1.26.0
pandas>=2.2.0

# UI компоненты
streamlit-extras>=0.3.5
plotly>=5.18.0

# Безопасность
pyjwt>=2.8.0
cryptography>=42.0.0

# Разработка
black>=23.12.0
pytest>=7.4.0